+++ configure.py
@@ -1180,6 +1180,7 @@ scylla_core = (['message/messaging_service.cc',
                 'mutation_writer/partition_based_splitting_writer.cc',
                 'mutation_writer/token_group_based_splitting_writer.cc',
                 'mutation_writer/feed_writers.cc',
+                'lang/manager.cc',
                 'lang/lua.cc',
                 'lang/wasm.cc',
                 'lang/wasm_alien_thread_runner.cc',
